requirements.txt
fastapi
uvicorn
python-dotenv
pandas
pydantic
matplotlib
seaborn
base64
itertools